TZ-9-Done CheckStyle(Rename server)
