+++ src/HttpTaskServer.java
@@ -1,7 +1,7 @@
 import com.sun.net.httpserver.HttpServer;
 
 import manager.*;
-import Server.*;
+import server.*;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
