+++ src/server/HistoryHandler.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
@@ -7,11 +7,11 @@ import manager.TaskManager;
 
 import java.io.IOException;
 
-public class PrioritizedTasksHandler extends BaseHttpHandler implements HttpHandler {
+public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
     private final TaskManager taskManager;
     private final Gson gson;
 
-    public PrioritizedTasksHandler(TaskManager taskManager, Gson gson) {
+    public HistoryHandler(TaskManager taskManager, Gson gson) {
         this.taskManager = taskManager;
         this.gson = gson;
     }
@@ -19,13 +19,12 @@ public class PrioritizedTasksHandler extends BaseHttpHandler implements HttpHand
     @Override
     public void handle(HttpExchange exchange) throws IOException {
         try {
-            if ("GET".equals(exchange.getRequestMethod()) && "/prioritized".equals(exchange.getRequestURI().getPath())) {
-                sendText(exchange, gson.toJson(taskManager.getPrioritizedTasks()), 200);
+            if ("GET".equals(exchange.getRequestMethod()) && "/history".equals(exchange.getRequestURI().getPath())) {
+                sendText(exchange, gson.toJson(taskManager.getHistory()), 200);
             } else {
                 sendNotFound(exchange);
             }
         } catch (Exception e) {
-            System.err.println("Error handling request: " + e.getMessage());
             sendError(exchange, "An error occurred: " + e.getMessage());
         }
     }
