+++ src/server/TaskHandler.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
